chore(client): rename websocket utility
